# backend/app/openapi.yml
openapi: 3.0.1
info:
  version: 2.0.0
  title: Threat designer APIs
  description: Threat designer api
security:
  - LambdaAuthorizer: []
...
paths:
  "/threat-designer/status/{id}":
    get:
      summary: Fetch Threat modeling statuses
      description: Fetch Threat modeling status
      tags:
        - Security
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Threat model identifier
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
      x-amazon-apigateway-integration:
        payloadFormatVersion: "2.0"
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "${lambda_arn}"
        connectionType: "INTERNET"
        contentHandling: CONVERT_TO_TEXT
        passthroughBehavior: NEVER
    options:
      responses:
        "200":
          description: OK
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Credentials:
              schema:
                type: string
      security: []
      tags:
        - CORS (Options)
      x-amazon-apigateway-integration:
        payloadFormatVersion: "2.0"
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "${lambda_arn}"
        connectionType: "INTERNET"
        contentHandling: CONVERT_TO_TEXT
        passthroughBehavior: NEVER
  "/threat-designer/mcp/status/{id}":
    get:
      summary: Fetch Threat modeling statuses (MCP)
      description: |
        Fetch Threat modeling status using MCP API key authentication.
        This endpoint provides the same functionality as the JWT version.
      tags:
        - Security
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Threat model identifier
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
      x-amazon-apigateway-integration:
        payloadFormatVersion: "2.0"
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "${lambda_arn}"
        connectionType: "INTERNET"
        contentHandling: CONVERT_TO_TEXT
        passthroughBehavior: NEVER
  "/threat-designer/restore/{id}":
    put:
      summary: Restore previous threat model
      description: Restore previous threat model
      tags:
        - Security
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Threat model identifier
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreatModelResponse"
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
      x-amazon-apigateway-integration:
        payloadFormatVersion: "2.0"
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "${lambda_arn}"
        connectionType: "INTERNET"
        contentHandling: CONVERT_TO_TEXT
        passthroughBehavior: NEVER
    options:
      responses:
        "200":
          description: OK
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Credentials:
              schema:
                type: string
      security: []
      tags:
        - CORS (Options)
      x-amazon-apigateway-integration:
        payloadFormatVersion: "2.0"
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "${lambda_arn}"
        connectionType: "INTERNET"
        contentHandling: CONVERT_TO_TEXT
        passthroughBehavior: NEVER
  "/threat-designer/mcp/restore/{id}":
    put:
      summary: Restore previous threat model (MCP)
      description: |
        Restore previous threat model using MCP API key authentication.
        This endpoint provides the same functionality as the JWT version.
        Note: Response does not include is_owner and access_level fields (JWT-only fields).
      tags:
        - Security
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Threat model identifier
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreatModelResponse"
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
      x-amazon-apigateway-integration:
        payloadFormatVersion: "2.0"
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "${lambda_arn}"
        connectionType: "INTERNET"
        contentHandling: CONVERT_TO_TEXT
        passthroughBehavior: NEVER
  "/threat-designer/trail/{id}":
    get:
      summary: Fetch Threat modeling reasoning trail
      description: Fetch Threat modeling reasoning trail
      tags:
        - Security
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Threat model identifier
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrailResponse"
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
      x-amazon-apigateway-integration:
        payloadFormatVersion: "2.0"
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "${lambda_arn}"
        connectionType: "INTERNET"
        contentHandling: CONVERT_TO_TEXT
        passthroughBehavior: NEVER
    options:
      responses:
        "200":
          description: OK
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Credentials:
              schema:
                type: string
      security: []
      tags:
        - CORS (Options)
      x-amazon-apigateway-integration:
        payloadFormatVersion: "2.0"
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "${lambda_arn}"
        connectionType: "INTERNET"
        contentHandling: CONVERT_TO_TEXT
        passthroughBehavior: NEVER
...
  "/threat-designer/all":
    get:
      summary: Fetch all threat models for a given user
      description: |
        Fetch threat models for a given user with pagination support.

        Returns a paginated list of threat models (both owned and shared) sorted by creation date (newest first).
        Use the cursor from the response to fetch subsequent pages.
      tags:
        - Security
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            enum: [10, 20, 50, 100]
            default: 20
          description: Number of items per page (must be 10, 20, 50, or 100)
          example: 20
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: |
            Base64-encoded pagination cursor from previous response.
            Omit this parameter to fetch the first page.
          example: "eyJvd25lZCI6ImJhc2U2NCIsInNoYXJlZCI6ImJhc2U2NCIsImZpbHRlciI6ImFsbCJ9"
        - name: filter
          in: query
          required: false
          schema:
            type: string
            enum: [owned, shared, all]
            default: all
          description: |
            Filter results by ownership:
            - owned: Only threat models owned by the user
            - shared: Only threat models shared with the user
            - all: Both owned and shared threat models
          example: all
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogResponse"
              examples:
                first_page:
                  summary: First page of results
                  value:
                    catalogs:
                      - job_id: "550e8400-e29b-41d4-a716-446655440000"
                        title: "E-commerce Platform Threat Model"
                        state: "COMPLETE"
                        created_at: "2024-01-15T10:30:00Z"
                        owner: "user-123"
                        is_owner: true
                        access_level: "OWNER"
                      - job_id: "550e8400-e29b-41d4-a716-446655440001"
                        title: "API Gateway Security Analysis"
                        state: "COMPLETE"
                        created_at: "2024-01-14T15:20:00Z"
                        owner: "user-456"
                        is_owner: false
                        access_level: "EDIT"
                        shared_by: "user-456"
                    pagination:
                      hasNextPage: true
                      cursor: "eyJvd25lZCI6ImJhc2U2NCIsInNoYXJlZCI6ImJhc2U2NCIsImZpbHRlciI6ImFsbCJ9"
                      totalReturned: 20
                last_page:
                  summary: Last page of results
                  value:
                    catalogs:
                      - job_id: "550e8400-e29b-41d4-a716-446655440099"
                        title: "Mobile App Security Review"
                        state: "COMPLETE"
                        created_at: "2024-01-01T09:00:00Z"
                        owner: "user-123"
                        is_owner: true
                        access_level: "OWNER"
                    pagination:
                      hasNextPage: false
                      cursor: null
                      totalReturned: 5
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
      x-amazon-apigateway-integration:
        payloadFormatVersion: "2.0"
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "${lambda_arn}"
        connectionType: "INTERNET"
        contentHandling: CONVERT_TO_TEXT
        passthroughBehavior: NEVER
    options:
      responses:
        "200":
          description: OK
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Credentials:
              schema:
                type: string
      security: []
      tags:
        - CORS (Options)
      x-amazon-apigateway-integration:
        payloadFormatVersion: "2.0"
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "${lambda_arn}"
        connectionType: "INTERNET"
        contentHandling: CONVERT_TO_TEXT
        passthroughBehavior: NEVER
  "/threat-designer/mcp/all":
    get:
      summary: Fetch all threat models for a given user (MCP)
      description: |
        Fetch threat models for a given user with pagination support using MCP API key authentication.
        This endpoint provides the same functionality as the JWT version.

        Returns a paginated list of threat models (both owned and shared) sorted by creation date (newest first).
        Use the cursor from the response to fetch subsequent pages.

        Note: Response items do not include is_owner and access_level fields (JWT-only fields).
      tags:
        - Security
      security:
        - ApiKeyAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            enum: [10, 20, 50, 100]
            default: 20
          description: Number of items per page (must be 10, 20, 50, or 100)
          example: 20
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: |
            Base64-encoded pagination cursor from previous response.
            Omit this parameter to fetch the first page.
          example: "eyJvd25lZCI6ImJhc2U2NCIsInNoYXJlZCI6ImJhc2U2NCIsImZpbHRlciI6ImFsbCJ9"
        - name: filter
          in: query
          required: false
          schema:
            type: string
            enum: [owned, shared, all]
            default: all
          description: |
            Filter results by ownership:
            - owned: Only threat models owned by the user
            - shared: Only threat models shared with the user
            - all: Both owned and shared threat models
          example: all
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogResponse"
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
      x-amazon-apigateway-integration:
        payloadFormatVersion: "2.0"
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "${lambda_arn}"
        connectionType: "INTERNET"
        contentHandling: CONVERT_TO_TEXT
        passthroughBehavior: NEVER
  "/threat-designer":
    post:
      summary: Submit Threat modeling
      description: Submit Threat modeling
      tags:
        - Security
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateThreatModelRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateThreatModelResponse"
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
      x-amazon-apigateway-integration:
        payloadFormatVersion: "2.0"
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "${lambda_arn}"
        connectionType: "INTERNET"
        contentHandling: CONVERT_TO_TEXT
        passthroughBehavior: NEVER
    options:
      responses:
        "200":
          description: OK
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Credentials:
              schema:
                type: string
      security: []
      tags:
        - CORS (Options)
      x-amazon-apigateway-integration:
        payloadFormatVersion: "2.0"
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "${lambda_arn}"
        connectionType: "INTERNET"
        contentHandling: CONVERT_TO_TEXT
        passthroughBehavior: NEVER
  "/threat-designer/mcp":
    post:
      summary: Submit Threat modeling (MCP)
      description: |
        Submit Threat modeling using MCP API key authentication.
        This endpoint provides the same functionality as the JWT version.
      tags:
        - Security
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateThreatModelRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateThreatModelResponse"
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
      x-amazon-apigateway-integration:
        payloadFormatVersion: "2.0"
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "${lambda_arn}"
        connectionType: "INTERNET"
        contentHandling: CONVERT_TO_TEXT
        passthroughBehavior: NEVER
  "/threat-designer/download":
    post:
      summary: Generates download presigned-url
      description: Generates download presigned-url
      tags:
        - Security
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DownloadRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresignedUrlResponse"
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
      x-amazon-apigateway-integration:
        payloadFormatVersion: "2.0"
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "${lambda_arn}"
        connectionType: "INTERNET"
        contentHandling: CONVERT_TO_TEXT
        passthroughBehavior: NEVER
    options:
      responses:
        "200":
          description: OK
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Credentials:
              schema:
                type: string
      security: []
      tags:
        - CORS (Options)
      x-amazon-apigateway-integration:
        payloadFormatVersion: "2.0"
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "${lambda_arn}"
        connectionType: "INTERNET"
        contentHandling: CONVERT_TO_TEXT
        passthroughBehavior: NEVER
...
components:
  schemas:
...
    CreateThreatModelResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
          description: The threat model identifier
...
    StatusResponse:
      type: object
      required:
        - id
        - state
        - retry
        - session_id
      properties:
        id:
          type: string
          format: uuid
          description: The threat model identifier
        state:
          type: string
          enum: [START, PENDING, PROCESSING, COMPLETE, FAILED, Not Found]
          description: Current execution state
        retry:
          type: integer
          description: Reasoning boost level used for this execution
        session_id:
          type: string
          format: uuid
          description: Unique identifier for the agent execution session
        execution_owner:
          type: string
          format: uuid
          description: User ID who started this execution (optional)
        detail:
          type: string
          description: Additional detail message about the state (optional)
...
  responses:
    "400":
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
      description: Client Errors
      headers:
        Access-Control-Allow-Headers:
          schema:
            type: string
        Access-Control-Allow-Origin:
          schema:
            type: string
...
    "500":
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
      description: Server Errors
      headers:
        Access-Control-Allow-Headers:
          schema:
            type: string
        Access-Control-Allow-Origin:
          schema:
            type: string
  securitySchemes:
    LambdaAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: "custom"
      x-amazon-apigateway-authorizer:
        identitySource: "method.request.header.authorization"
        authorizerUri: ${authorizer_arn}
        authorizerResultTtlInSeconds: 0
        type: token
        authorizerPayloadFormatVersion: "1.0"
    ApiKeyAuth:
      type: apiKey
      name: x-api-key
      in: header